Untangle TURN permissions and channel bindings
As per RFC 5766 (section 8), if a ChannelBind request is accepted, a
permission for the peer's IP address must be installed or refreshed
implicitly:

| CreatePermission and ChannelBind requests may be freely intermixed on
| a permission.  A given permission may be initially installed and/or
| refreshed with a CreatePermission request, and then later refreshed
| with a ChannelBind request, or vice versa.

Therefore, don't expect the client to perform an explicit
CreatePermission request before binding a channel.

Also, when receiving traffic from a peer, don't ignore the peer's source
port while assigning the traffic to a channel.  Otherwise, the mapping
would go wrong in the case where multiple peers sit behind the same IP
address.

As per RFC 5766 (section 11.2), reject ChannelBind requests that attempt
to bind a channel to a peer that's already bound to another channel, and
reject requests that attempt to bind a new peer to an existing channel.

Closes #22.