Accept expired credentials for TURN refreshes
If the authentication function tags the returned password as 'expired',
let the behavior depend on the context: reject initial TURN allocation
but (if the credentials are valid otherwise) accept session refreshes.
Otherwise, the expiry timestamp of ephemeral credentials would impose a
hard limit on the call time, which may not be desired.